speed 수정

+++ Assignment2/Boat.cpp
@@ -18,7 +18,7 @@ namespace assignment2
 	unsigned int Boat::GetSailSpeed() const
 	{
 		double totalWeight = 0;
-		unsigned int result = 20;
+		int result = 20;
 
 		if (GetPassengersCount() > 0)
 		{
@@ -27,10 +27,10 @@ namespace assignment2
 				totalWeight += static_cast<double>(GetPassenger(i)->GetWeight());
 			}
 		}
-		result = static_cast<unsigned int>(800.0 - (totalWeight * 10.0) + 0.5);
+		result = static_cast<int>(800.0 - (totalWeight * 10.0) + 0.5);
 		if (20 < result)
 		{
-			return result;
+			return static_cast<unsigned int>(result);
 		}
 
 		return 20;
