+++ Assignment2/Boatplane.cpp
@@ -48,7 +48,7 @@ namespace assignment2
 	unsigned int Boatplane::GetSailSpeed() const
 	{
 		double totalWeight = 0;
-		unsigned int result = 0;
+		int result = 0;
 
 		if (GetPassengersCount() > 0)
 		{
@@ -58,11 +58,11 @@ namespace assignment2
 			}
 		}
 
-		result = static_cast<unsigned int>((800.0 - totalWeight * 1.7) + 0.5);
+		result = static_cast<int>((800.0 - totalWeight * 1.7) + 0.5);
 
 		if (20 < result)
 		{
-			return result;
+			return static_cast<unsigned int>(result);
 		}
 
 		return 20;
